Added compatibility for 1.0.2 as well as 1.0.3

+++ JET_ProfileEditor/Character.cs
@@ -267,7 +267,7 @@ namespace JET_ProfileEditor
                     public class Character_Inventory_Item_Upd_MedKit
                     {
                         [JsonProperty("HpResource")]
-                        public int HpResource { get; set; }
+                        public double HpResource { get; set; }
                     }
 
                     public class Character_Inventory_Item_Upd_FireMode
